feat(calculator): 연산 객체 분리, 새로운 로직, 예외 클래스 추가,

+++ src/calculator/Calculator.java
@@ -14,7 +14,7 @@ public class Calculator {
         this.resultHistory = new ArrayList<>();
     }
 
-    public double calculate(int a, int b, char operator) {
+    public double calculate(double a, double b, char operator) {
         double result;
 
         if (operator == '+') {
